Refactor image resources and update UI dimensions
Updated `EZLogger.vbproj` to replace image files in the `Views` directory with resource files in the `Resources` directory. Increased the size of `ElementHost1` and the `ClientSize` of the `FileSaverHost` form from 516 to 615 pixels. Modified image source paths in `FileSaverControl.xaml` to reflect the new resource locations. Deleted several binary image files from the project.

Refactor image resources and update UI dimensions

Updated `EZLogger.vbproj` to replace image files in the `Views` directory with resource files in the `Resources` directory. Increased the height of `ElementHost1` and the `ClientSize` of the `FileSaverHost` form from 516 to 615 pixels. Modified image source paths in `FileSaverControl.xaml` to reflect the new resource locations. Deleted obsolete binary image files.